refactor(sql_core): split sql_core between different files
explode sql_core in 3 file : .1 sql_core that contains all basic sql logic (database export, request ...) .2 sql_main contains parse_sql_generate_command_from_list and other sub components .3 sql_command_constructor contains SQLCommandFormater

#4
