+++ job_scrapper/main.py
@@ -396,7 +396,7 @@ def request(
     "--display",
     is_flag=True,
     help="Output the result inside the terminal. If --result-file (-f) is not used and "
-    "--no-sql-export (-q) is used, this flag"
+    "--no-sql_core-export (-q) is used, this flag"
     "is turned on by default.",
 )
 @click.option(
@@ -408,7 +408,7 @@ def request(
 )
 @click.option(
     "-q",
-    "--no-sql-export",
+    "--no-sql_core-export",
     is_flag=True,
     help="The result will not be exported inside a database in the workdir.",
 )
