+++ job_scrapper/scrapper_skeleton/request_core.py
@@ -23,12 +23,12 @@ from selenium.webdriver.remote.webelement import WebElement
 from selenium.webdriver.support import expected_conditions as EC
 from selenium.webdriver.support.ui import WebDriverWait
 
-from .sql_core import ScrapperSQLightCore
+from .sql_core.sql_main import ScrapperSQLightRunner
 
 
-class ScrapperRequestCore(ScrapperSQLightCore):
+class ScrapperRequestCore(ScrapperSQLightRunner):
     """
-    Specialisation of ScrapperSQLightCore. Add the ability to fetch
+    Specialisation of ScrapperSQLightRunner. Add the ability to fetch
     website's jobs, count keywords occurrences inside the description
     of an offer and compute distances between two localisation
     (e.g. job' address and a city)
